Break types into files
